requirements.txt
pydantic>=2.4.0
pyyaml>=6.0.1
aiosqlite>=0.19.0
chromadb>=0.4.15
fastapi>=0.104.0
uvicorn>=0.23.2
websockets>=11.0.3
ollama>=0.1.7
